Added Files and edited eventpage
I added the files for bootstrap and jquery so y'all can use it in your
pages. I also turn the event page into a popup modal, but I didn't want
to edit the calendar page so I left it as its own page for now. Just
click the title to see it.