Making the events dynamic
I've added php functionality to the create event and view event page,
and added a submit event page that is just a redirect. It works as a
proof of concept, but I'll need to update it because i'm still trying to
figure out foreign key stuff. It won't work unless you've installed php,
apache and mysql, and have a mysql database called "LTUBillboard" with a
table called "LTUEvents" and username/password of root/root. The table
is also set up with columns: (evtPrivate, evtName,
evtBuildingRoom,evtCategory,evtDate,evtTime,evtDesc,evtOrgId)