MYSQL table creation
Here's what I think the mysql tables could be. The main problem is that
I'm not sure how the user_account table should be, since they need to be
able to reference many events and organizations. I was thinking just
have be text, with comma separating the different ids of
events/organizations