new format
I've combined mine and Kat's format to create one that I'm pretty sure
will work. It uses two extra tables that can be used to search for
events that a user signed up for and organizations that a user has
subscribed to. I'm going to work now on updating my php to do all this.