work on header
Made it so the header can be included in any file with the php required
function. It should go after any of the custom javascript on the page,
but before any content that deals with the page itself. There should
also be no ending head tag or starting body tag in the file that it's
going to be included in. See fcIndex.php for example.