changed included header and am working on making calendar dynamic
I had to update the required header and everything that used it because
it couldn't have a doc.ready function. I'm trying to get the events to
show a modal on the index page but it's not working. fcIndex is what it
was before and fcIndex2 is what i'm working on.