Files for Admin Page
The admin page is still a work-in-progress, but I wanted to show you what I've done so far. I was finally able to update the database when the administrator accepts or declines a request. For the entire page to work, you need to add two additional columns.

In ltuevents add:  evt_visible TINYINT(1) NOT NULL,
In ltuorganization add:   org_accepted TINYINT(1) NOT NULL

Events/Organizations will show on the request table when these parameters are set to 0. I used ajax to send asynchronous xhttprequests to the database. This way the table updates after an action without the user needing to refresh the page.
 
Basically,  admin.php uses findMax.php to find the maximum number of "pages" that can be displayed in the requests table. buildRequestTable.php will fill this table with 20 requests at a time.  accpetOrDecline.php handles updating or deleting a database entry and upcomingEvents.php shows the next 5 upcoming events.