Added ability for user to create events
Users can now create events for just them to view in their board. The
org id is set to the negative of whatever their user id is, they event
is auto set to visible, and org name is set to "User Created". I've also
updated the filters to take these into account and added a filter for
just those events.